% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with three columns:
- `v1`: starting node of an edge
- `v2`: ending node of an edge
- `w`: weight of the edge (numeric, non-negative)}

\item{init_node}{A numeric scalar representing the starting node.}
}
\value{
A numeric vector where each entry gives the shortest distance from `init_node`
to the corresponding node in the graph.
}
\description{
Computes the shortest paths from an initial node to all other nodes in a weighted graph.
}
\examples{
wiki_graph <- data.frame(
  v1 = c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2 = c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w  = c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
