% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Dataset}
\format{
A data.frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Integer. Starting node of an edge.}
  \item{v2}{Integer. Ending node of an edge.}
  \item{w}{Numeric. Weight of the edge.}
}
}
\source{
Wikipedia: Dijkstra's algorithm, available at
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
A sample graph represented as a data.frame with edges and weights.
This graph is used to demonstrate Dijkstra's algorithm.
}
\keyword{datasets}
